Create src/sampler package
